Docs WIP, updated screenshotAll() for something more open, tests updated
