+++ src/botmation/actions/output.ts
@@ -12,22 +12,21 @@ import { getFileUrl } from '../helpers/assets'
  */
 export const screenshot = (fileName: string): BotAction => async(page: Page, options) => {
   const fileUrl = getFileUrl(options.screenshots_directory, options, fileName) + '.png'
-  // console.log('[screenshot] fileUrl = ' + fileUrl)
 
   await page.screenshot({path: fileUrl})
 }
 
 /**
  * @description    given a list of websites, the bot will visit each, wait for them to load, then take a screenshot to save in the escreenshot directory
- * @param sites ['example.com', 'whatever.com']
- * @example   screenshotAll('google.com', 'twitter.com')
+ * @param urls ['https://example.com', 'http://whatever.com']
+ * @example   screenshotAll('https://google.com', 'https://twitter.com')
  * @experimental
  * @request   add ability like via a closure, to customize the filename for easier reuse in a cycle (like ability to timestamp the file etc)
  */
-export const screenshotAll = (...sites: string[]): BotAction => async(page: Page, options, ...injects:[]) =>
-  forAll(sites)(
-    (siteName) => ([
-      goTo('https://' + siteName),
-      screenshot(siteName)
+export const screenshotAll = (...urls: string[]): BotAction => async(page: Page, options) =>
+  await forAll(urls)(
+    (url) => ([
+      goTo(url),
+      screenshot(url.replace(/[^a-zA-Z]/g, '_')) // filenames are created from urls by replacing nonsafe characters with underscore
     ])
-  )(page, options, ...injects)
+  )(page, options)
