+++ src/tests/README.md
@@ -1,7 +1,21 @@
 <h1>Botmation: Tests</h1>
 
-Bot Actions are tested. Class, Factory are tested.
+Given this is a library, testing is very important. 
 
-Missing test is on `wait()`
+It's all done by Jest and found in this directory `/src/test`.
 
-CLI scripts.... more docs coming soon
+Unit-Tests
+------------
+Every Bot Action Factory, except `wait()`, is covered with unit-testing. 
+
+Integration Tests
+-----------------
+Every Bot Action that uses one or more of Puppeteer's Page's public methods, has an Integration test included.
+
+End-to-End Tests
+----------------
+The class and main factory have E2E tests, running in puppeteer. This is to be sure the class and factory are always working.
+
+Testing Scaffolding
+-------------------
+The `/src/tests/botmation` directory follows a 1:1 mapping with the `/src/botmation` directory, as to where each files' tests are.
